+++ cuda_core/cuda/core/__init__.py
@@ -2,9 +2,4 @@
 #
 # SPDX-License-Identifier: LicenseRef-NVIDIA-SOFTWARE-LICENSE
 
-from cuda.core._device import Device
-from cuda.core._event import EventOptions
-from cuda.core._launcher import LaunchConfig, launch
-from cuda.core._program import Program
-from cuda.core._stream import Stream, StreamOptions
 from cuda.core._version import __version__
