+++ cuda_core/cuda/core/experimental/_context.py
@@ -5,7 +5,7 @@
 from dataclasses import dataclass
 
 from cuda import cuda, cudart
-from cuda.core._utils import handle_return
+from cuda.core.experimental._utils import handle_return
 
 
 @dataclass
