+++ cuda_core/cuda/core/experimental/_launcher.py
@@ -8,11 +8,11 @@ from typing import Optional, Union
 import numpy as np
 
 from cuda import cuda, cudart
-from cuda.core._kernel_arg_handler import ParamHolder
-from cuda.core._memory import Buffer
-from cuda.core._module import Kernel
-from cuda.core._stream import Stream
-from cuda.core._utils import CUDAError, check_or_create_options, handle_return
+from cuda.core.experimental._kernel_arg_handler import ParamHolder
+from cuda.core.experimental._memory import Buffer
+from cuda.core.experimental._module import Kernel
+from cuda.core.experimental._stream import Stream
+from cuda.core.experimental._utils import CUDAError, check_or_create_options, handle_return
 
 
 @dataclass
