+++ cuda_core/cuda/core/experimental/_module.py
@@ -3,7 +3,7 @@
 # SPDX-License-Identifier: LicenseRef-NVIDIA-SOFTWARE-LICENSE
 
 from cuda import cuda, cudart
-from cuda.core._utils import handle_return
+from cuda.core.experimental._utils import handle_return
 
 
 _backend = {
