+++ cuda_core/examples/saxpy.py
@@ -4,9 +4,9 @@
 
 import sys
 
-from cuda.core import Device
-from cuda.core import LaunchConfig, launch
-from cuda.core import Program
+from cuda.core.experimental import Device
+from cuda.core.experimental import LaunchConfig, launch
+from cuda.core.experimental import Program
 
 import cupy as cp
